Add range operator
Currently the expression '....' is seen as valid. This is not the case, as it'd concatenate a range starting with -Infinity with one ending with Infinity-1. This will always yield the same result as a simple '..' expression. I don't know if I can find a way to make this cause an error on a first pass, but on a second pass it may be possible. In that case I'll have to write it in JS, however.