Add type definitions
These definitions have been created in preparation for a compilation pass, and can be used to infer types from expressions. The method typesOfValue works, and returns a list of the most specific matching types (i.e. a negative integer such as -5 will return TYPES.Negative and TYPES.Integer, but not TYPES.Number or TYPES.Any, since those are implied). This could cause some issues in the future with the Char type specifically, as it can be counted as both a number subtype or a string subtype, so perhaps those deserve to be separate types to avoid strange outcomes when doing arithmetic with them (i.e. Char+Char is... String with length 2? Or does it stay a Char, simply summing the code points?)